+++ psutil/__init__.py
@@ -15,8 +15,10 @@ sensors) in Python. Supported platforms:
  - Sun Solaris
  - AIX
 
-Supported Python versions are cPython 3.6+ and PyPy.
+Supported Python versions are cPython 3.7+ and PyPy.
 """
+
+from __future__ import annotations
 
 import collections
 import contextlib
@@ -30,6 +32,7 @@ import sys
 import threading
 import time
 import warnings
+from typing import TYPE_CHECKING as _TYPE_CHECKING
 
 try:
     import pwd
@@ -61,6 +64,48 @@ from ._enums import BatteryTime
 from ._enums import ConnectionStatus
 from ._enums import NicDuplex
 from ._enums import ProcessStatus
+
+if _TYPE_CHECKING:
+    from typing import Any
+    from typing import Callable
+    from typing import Generator
+    from typing import Iterator
+
+    from ._ntuples import pconn
+    from ._ntuples import pcputimes
+    from ._ntuples import pctxsw
+    from ._ntuples import pfootprint
+    from ._ntuples import pfullmem
+    from ._ntuples import pgids
+    from ._ntuples import pheap
+    from ._ntuples import pio
+    from ._ntuples import pionice
+    from ._ntuples import pmem
+    from ._ntuples import pmem_ex
+    from ._ntuples import pmmap_ext
+    from ._ntuples import pmmap_grouped
+    from ._ntuples import popenfile
+    from ._ntuples import ppagefaults
+    from ._ntuples import pthread
+    from ._ntuples import puids
+    from ._ntuples import sbattery
+    from ._ntuples import sconn
+    from ._ntuples import scpufreq
+    from ._ntuples import scpustats
+    from ._ntuples import scputimes
+    from ._ntuples import sdiskio
+    from ._ntuples import sdiskpart
+    from ._ntuples import sdiskusage
+    from ._ntuples import sfan
+    from ._ntuples import shwtemp
+    from ._ntuples import snetio
+    from ._ntuples import snicaddr
+    from ._ntuples import snicstats
+    from ._ntuples import sswap
+    from ._ntuples import suser
+    from ._ntuples import svmem
+    from ._pswindows import WindowsService
+
 
 if LINUX:
     # This is public API and it will be retrieved from _pslinux.py
@@ -271,7 +316,7 @@ class Process:
     is_running() before querying the process.
     """
 
-    def __init__(self, pid=None):
+    def __init__(self, pid: int | None = None) -> None:
         self._init(pid)
 
     def _init(self, pid, _ignore_nsp=False):
@@ -430,14 +475,14 @@ class Process:
             raise NoSuchProcess(self.pid, self._name, msg=msg)
 
     @property
-    def pid(self):
+    def pid(self) -> int:
         """The process PID."""
         return self._pid
 
     # --- utility methods
 
     @contextlib.contextmanager
-    def oneshot(self):
+    def oneshot(self) -> Generator[None, None, None]:
         """Utility context manager which considerably speeds up the
         retrieval of multiple process information at the same time.
 
@@ -503,7 +548,9 @@ class Process:
                         self.uids.cache_deactivate(self)
                     self._proc.oneshot_exit()
 
-    def as_dict(self, attrs=None, ad_value=None):
+    def as_dict(
+        self, attrs: list[str] | None = None, ad_value: Any = None
+    ) -> dict[str, Any]:
         """Utility method returning process information as a
         hashable dictionary.
         If *attrs* is specified it must be a list of strings
@@ -550,7 +597,7 @@ class Process:
                 retdict[name] = ret
         return retdict
 
-    def parent(self):
+    def parent(self) -> Process | None:
         """Return the parent process as a Process object pre-emptively
         checking whether PID has been reused.
         If no parent is known return None.
@@ -572,7 +619,7 @@ class Process:
             except NoSuchProcess:
                 pass
 
-    def parents(self):
+    def parents(self) -> list[Process]:
         """Return the parents of this process as a list of Process
         instances. If no parents are known return an empty list.
         """
@@ -583,7 +630,7 @@ class Process:
             proc = proc.parent()
         return parents
 
-    def is_running(self):
+    def is_running(self) -> bool:
         """Return whether this process is running.
 
         It also checks if PID has been reused by another process, in
@@ -613,7 +660,7 @@ class Process:
     # --- actual API
 
     @memoize_when_activated
-    def ppid(self):
+    def ppid(self) -> int:
         """The process parent PID.
         On Windows the return value is cached after first call.
         """
@@ -631,7 +678,7 @@ class Process:
             self._ppid = self._ppid or self._proc.ppid()
             return self._ppid
 
-    def name(self):
+    def name(self) -> str:
         """The process name. The return value is cached after first call."""
         # Process name is only cached on Windows as on POSIX it may
         # change, see:
@@ -662,7 +709,7 @@ class Process:
         self._proc._name = name
         return name
 
-    def exe(self):
+    def exe(self) -> str:
         """The process executable as an absolute path.
         May also be an empty string.
         The return value is cached after first call.
@@ -704,18 +751,18 @@ class Process:
                 self._exe = exe
         return self._exe
 
-    def cmdline(self):
+    def cmdline(self) -> list[str]:
         """The command line this process has been called with."""
         return self._proc.cmdline()
 
-    def status(self):
+    def status(self) -> ProcessStatus | str:
         """The process current status as a STATUS_* constant."""
         try:
             return self._proc.status()
         except ZombieProcess:
             return ProcessStatus.STATUS_ZOMBIE
 
-    def username(self):
+    def username(self) -> str:
         """The name of the user that owns the process.
         On UNIX this is calculated by using *real* process uid.
         """
@@ -733,7 +780,7 @@ class Process:
         else:
             return self._proc.username()
 
-    def create_time(self):
+    def create_time(self) -> float:
         """The process creation time as a floating point number
         expressed in seconds since the epoch (seconds since January 1,
         1970, at midnight UTC). The return value, which is cached after
@@ -745,11 +792,11 @@ class Process:
             self._create_time = self._proc.create_time()
         return self._create_time
 
-    def cwd(self):
+    def cwd(self) -> str:
         """Process current working directory as an absolute path."""
         return self._proc.cwd()
 
-    def nice(self, value=None):
+    def nice(self, value: int | None = None) -> int | None:
         """Get or set process niceness (priority)."""
         if value is None:
             return self._proc.nice_get()
@@ -760,25 +807,25 @@ class Process:
     if POSIX:
 
         @memoize_when_activated
-        def uids(self):
+        def uids(self) -> puids:
             """Return process UIDs as a (real, effective, saved)
             namedtuple.
             """
             return self._proc.uids()
 
-        def gids(self):
+        def gids(self) -> pgids:
             """Return process GIDs as a (real, effective, saved)
             namedtuple.
             """
             return self._proc.gids()
 
-        def terminal(self):
+        def terminal(self) -> str | None:
             """The terminal associated with this process, if any,
             else None.
             """
             return self._proc.terminal()
 
-        def num_fds(self):
+        def num_fds(self) -> int:
             """Return the number of file descriptors opened by this
             process (POSIX only).
             """
@@ -787,7 +834,7 @@ class Process:
     # Linux, BSD, AIX and Windows only
     if hasattr(_psplatform.Process, "io_counters"):
 
-        def io_counters(self):
+        def io_counters(self) -> pio:
             """Return process I/O statistics as a
             (read_count, write_count, read_bytes, write_bytes)
             namedtuple.
@@ -799,7 +846,9 @@ class Process:
     # Linux and Windows
     if hasattr(_psplatform.Process, "ionice_get"):
 
-        def ionice(self, ioclass=None, value=None):
+        def ionice(
+            self, ioclass: int | None = None, value: int | None = None
+        ) -> pionice | ProcessIOPriority | None:
             """Get or set process I/O niceness (priority).
 
             On Linux *ioclass* is one of the IOPRIO_CLASS_* constants.
@@ -823,7 +872,11 @@ class Process:
     # Linux / FreeBSD only
     if hasattr(_psplatform.Process, "rlimit"):
 
-        def rlimit(self, resource, limits=None):
+        def rlimit(
+            self,
+            resource: int,
+            limits: tuple[int, int] | None = None,
+        ) -> tuple[int, int] | None:
             """Get or set process resource limits as a (soft, hard)
             tuple.
 
@@ -840,7 +893,9 @@ class Process:
     # Windows, Linux and FreeBSD only
     if hasattr(_psplatform.Process, "cpu_affinity_get"):
 
-        def cpu_affinity(self, cpus=None):
+        def cpu_affinity(
+            self, cpus: list[int] | None = None
+        ) -> list[int] | None:
             """Get or set process CPU affinity.
             If specified, *cpus* must be a list of CPUs for which you
             want to set the affinity (e.g. [0, 1]).
@@ -862,7 +917,7 @@ class Process:
     # Linux, FreeBSD, SunOS
     if hasattr(_psplatform.Process, "cpu_num"):
 
-        def cpu_num(self):
+        def cpu_num(self) -> int:
             """Return what CPU this process is currently running on.
             The returned number should be <= psutil.cpu_count()
             and <= len(psutil.cpu_percent(percpu=True)).
@@ -875,7 +930,7 @@ class Process:
     # All platforms has it, but maybe not in the future.
     if hasattr(_psplatform.Process, "environ"):
 
-        def environ(self):
+        def environ(self) -> dict[str, str]:
             """The environment variables of the process as a dict.  Note: this
             might not reflect changes made after the process started.
             """
@@ -883,25 +938,25 @@ class Process:
 
     if WINDOWS:
 
-        def num_handles(self):
+        def num_handles(self) -> int:
             """Return the number of handles opened by this process
             (Windows only).
             """
             return self._proc.num_handles()
 
-    def num_ctx_switches(self):
+    def num_ctx_switches(self) -> pctxsw:
         """Return the number of voluntary and involuntary context
         switches performed by this process.
         """
         return self._proc.num_ctx_switches()
 
-    def num_threads(self):
+    def num_threads(self) -> int:
         """Return the number of threads used by this process."""
         return self._proc.num_threads()
 
     if hasattr(_psplatform.Process, "threads"):
 
-        def threads(self):
+        def threads(self) -> list[pthread]:
             """Return threads opened by process as a list of
             (id, user_time, system_time) namedtuples representing
             thread id and thread CPU times (user/system).
@@ -909,7 +964,7 @@ class Process:
             """
             return self._proc.threads()
 
-    def children(self, recursive=False):
+    def children(self, recursive: bool = False) -> list[Process]:
         """Return the children of this process as a list of Process
         instances, pre-emptively checking whether PID has been reused.
         If *recursive* is True return all the parent descendants.
@@ -983,7 +1038,7 @@ class Process:
                         pass
         return ret
 
-    def cpu_percent(self, interval=None):
+    def cpu_percent(self, interval: float | None = None) -> float:
         """Return a float representing the current process CPU
         utilization as a percentage.
 
@@ -1077,7 +1132,7 @@ class Process:
             return round(single_cpu_percent, 1)
 
     @memoize_when_activated
-    def cpu_times(self):
+    def cpu_times(self) -> pcputimes:
         """Return a (user, system, children_user, children_system)
         namedtuple representing the accumulated process time, in
         seconds.
@@ -1088,7 +1143,7 @@ class Process:
         return self._proc.cpu_times()
 
     @memoize_when_activated
-    def memory_info(self):
+    def memory_info(self) -> pmem:
         """Return a namedtuple with variable fields depending on the
         platform, representing memory information about the process.
 
@@ -1098,7 +1153,8 @@ class Process:
         """
         return self._proc.memory_info()
 
-    def memory_info_ex(self):
+    @memoize_when_activated
+    def memory_info_ex(self) -> pmem_ex:
         """Return a namedtuple extending memory_info() with extra
         metrics.
 
@@ -1113,7 +1169,7 @@ class Process:
     # Linux, macOS, Windows
     if hasattr(_psplatform.Process, "memory_footprint"):
 
-        def memory_footprint(self):
+        def memory_footprint(self) -> pfootprint:
             """Return a named tuple with USS, PSS and swap memory
             metrics. These provide a better representation of
             actual process memory usage.
@@ -1129,7 +1185,7 @@ class Process:
             return self._proc.memory_footprint()
 
     # DEPRECATED
-    def memory_full_info(self):
+    def memory_full_info(self) -> pfullmem:
         """Return the same information as memory_info() plus
         memory_footprint() in a single named tuple.
 
@@ -1145,7 +1201,7 @@ class Process:
             return _ntp.pfullmem(*basic_mem + fp)
         return _ntp.pfullmem(*basic_mem)
 
-    def memory_percent(self, memtype="rss"):
+    def memory_percent(self, memtype: str = "rss") -> float:
         """Compare process memory to total physical system memory and
         calculate process memory utilization as a percentage.
         *memtype* argument is a string that dictates what type of
@@ -1194,7 +1250,9 @@ class Process:
 
     if hasattr(_psplatform.Process, "memory_maps"):
 
-        def memory_maps(self, grouped=True):
+        def memory_maps(
+            self, grouped: bool = True
+        ) -> list[pmmap_grouped] | list[pmmap_ext]:
             """Return process' mapped memory regions as a list of namedtuples
             whose fields are variable depending on the platform.
 
@@ -1219,7 +1277,7 @@ class Process:
             else:
                 return [_ntp.pmmap_ext(*x) for x in it]
 
-    def page_faults(self):
+    def page_faults(self) -> ppagefaults:
         """Return the number of page faults for this process as a
         (minor, major) namedtuple.
 
@@ -1237,14 +1295,14 @@ class Process:
         """
         return self._proc.page_faults()
 
-    def open_files(self):
+    def open_files(self) -> list[popenfile]:
         """Return files opened by process as a list of
         (path, fd) namedtuples including the absolute file name
         and file descriptor number.
         """
         return self._proc.open_files()
 
-    def net_connections(self, kind='inet'):
+    def net_connections(self, kind: str = "inet") -> list[pconn]:
         """Return socket connections opened by process as a list of
         (fd, family, type, laddr, raddr, status) namedtuples.
         The *kind* parameter filters for connections that match the
@@ -1270,7 +1328,7 @@ class Process:
         return self._proc.net_connections(kind)
 
     @_common.deprecated_method(replacement="net_connections")
-    def connections(self, kind="inet"):
+    def connections(self, kind="inet") -> list[pconn]:
         return self.net_connections(kind=kind)
 
     # --- signals
@@ -1302,7 +1360,7 @@ class Process:
             except PermissionError as err:
                 raise AccessDenied(pid, name) from err
 
-    def send_signal(self, sig):
+    def send_signal(self, sig: int) -> None:
         """Send a signal *sig* to process pre-emptively checking
         whether PID has been reused (see signal module constants) .
         On Windows only SIGTERM is valid and is treated as an alias
@@ -1317,7 +1375,7 @@ class Process:
                 raise NoSuchProcess(self.pid, self._name, msg=msg)
             self._proc.send_signal(sig)
 
-    def suspend(self):
+    def suspend(self) -> None:
         """Suspend process execution with SIGSTOP pre-emptively checking
         whether PID has been reused.
         On Windows this has the effect of suspending all process threads.
@@ -1328,7 +1386,7 @@ class Process:
             self._raise_if_pid_reused()
             self._proc.suspend()
 
-    def resume(self):
+    def resume(self) -> None:
         """Resume process execution with SIGCONT pre-emptively checking
         whether PID has been reused.
         On Windows this has the effect of resuming all process threads.
@@ -1339,7 +1397,7 @@ class Process:
             self._raise_if_pid_reused()
             self._proc.resume()
 
-    def terminate(self):
+    def terminate(self) -> None:
         """Terminate the process with SIGTERM pre-emptively checking
         whether PID has been reused.
         On Windows this is an alias for kill().
@@ -1350,7 +1408,7 @@ class Process:
             self._raise_if_pid_reused()
             self._proc.kill()
 
-    def kill(self):
+    def kill(self) -> None:
         """Kill the current process with SIGKILL pre-emptively checking
         whether PID has been reused.
         """
@@ -1360,7 +1418,7 @@ class Process:
             self._raise_if_pid_reused()
             self._proc.kill()
 
-    def wait(self, timeout=None):
+    def wait(self, timeout: float | None = None) -> int | None:
         """Wait for process to terminate, and if process is a children
         of os.getpid(), also return its exit code, else None.
         On Windows there's no such limitation (exit code is always
@@ -1460,7 +1518,7 @@ class Popen(Process):
     def __dir__(self):
         return sorted(set(dir(Popen) + dir(subprocess.Popen)))
 
-    def __enter__(self):
+    def __enter__(self) -> Popen:
         if hasattr(self.__subproc, '__enter__'):
             self.__subproc.__enter__()
         return self
@@ -1491,7 +1549,7 @@ class Popen(Process):
                 msg = f"{self.__class__!r} has no attribute {name!r}"
                 raise AttributeError(msg) from None
 
-    def wait(self, timeout=None):
+    def wait(self, timeout: float | None = None) -> int | None:
         if self.__subproc.returncode is not None:
             return self.__subproc.returncode
         ret = super().wait(timeout)
@@ -1504,7 +1562,7 @@ class Popen(Process):
 # =====================================================================
 
 
-def pids():
+def pids() -> list[int]:
     """Return a list of current running PIDs."""
     global _LOWEST_PID
     ret = sorted(_psplatform.pids())
@@ -1512,7 +1570,7 @@ def pids():
     return ret
 
 
-def pid_exists(pid):
+def pid_exists(pid: int) -> bool:
     """Return True if given PID exists in the current process list.
     This is faster than doing "pid in psutil.pids()" and
     should be preferred.
@@ -1534,7 +1592,9 @@ _pmap = {}
 _pids_reused = set()
 
 
-def process_iter(attrs=None, ad_value=None):
+def process_iter(
+    attrs: list[str] | None = None, ad_value: Any = None
+) -> Iterator[Process]:
     """Return a generator yielding a Process instance for all
     running processes.
 
@@ -1592,7 +1652,11 @@ process_iter.cache_clear = lambda: _pmap.clear()  # noqa: PLW0108
 process_iter.cache_clear.__doc__ = "Clear process_iter() internal cache."
 
 
-def wait_procs(procs, timeout=None, callback=None):
+def wait_procs(
+    procs: list[Process],
+    timeout: float | None = None,
+    callback: Callable[[Process], None] | None = None,
+) -> tuple[list[Process], list[Process]]:
     """Convenience function which waits for a list of processes to
     terminate.
 
@@ -1689,7 +1753,7 @@ def wait_procs(procs, timeout=None, callback=None):
 # =====================================================================
 
 
-def cpu_count(logical=True):
+def cpu_count(logical: bool = True) -> int | None:
     """Return the number of logical CPUs in the system (same as
     os.cpu_count()).
 
@@ -1712,7 +1776,7 @@ def cpu_count(logical=True):
     return ret
 
 
-def cpu_times(percpu=False):
+def cpu_times(percpu: bool = False) -> scputimes | list[scputimes]:
     """Return system-wide CPU times as a namedtuple.
     Every CPU time represents the seconds the CPU has spent in the
     given mode. The namedtuple's fields availability varies depending on the
@@ -1808,7 +1872,9 @@ def _cpu_times_deltas(t1, t2):
     return _ntp.scputimes(*field_deltas)
 
 
-def cpu_percent(interval=None, percpu=False):
+def cpu_percent(
+    interval: float | None = None, percpu: bool = False
+) -> float | list[float]:
     """Return a float representing the current system-wide CPU
     utilization as a percentage.
 
@@ -1890,7 +1956,9 @@ _last_cpu_times_2 = _last_cpu_times.copy()
 _last_per_cpu_times_2 = _last_per_cpu_times.copy()
 
 
-def cpu_times_percent(interval=None, percpu=False):
+def cpu_times_percent(
+    interval: float | None = None, percpu: bool = False
+) -> scputimes | list[scputimes]:
     """Same as cpu_percent() but provides utilization percentages
     for each specific CPU time as is returned by cpu_times().
     For instance, on Linux we'll get:
@@ -1949,14 +2017,14 @@ def cpu_times_percent(interval=None, percpu=False):
         return ret
 
 
-def cpu_stats():
+def cpu_stats() -> scpustats:
     """Return CPU statistics."""
     return _psplatform.cpu_stats()
 
 
 if hasattr(_psplatform, "cpu_freq"):
 
-    def cpu_freq(percpu=False):
+    def cpu_freq(percpu: bool = False) -> scpufreq | list[scpufreq] | None:
         """Return CPU frequency as a namedtuple including current,
         min and max frequency expressed in Mhz.
 
@@ -1999,7 +2067,7 @@ if hasattr(_psplatform, "cpu_freq"):
     __all__.append("cpu_freq")
 
 
-def getloadavg():
+def getloadavg() -> tuple[float, float, float]:
     """Return the average system load over the last 1, 5 and 15 minutes
     as a tuple. On Windows this is emulated by using a Windows API that
     spawns a thread which keeps running in background and updates
@@ -2008,7 +2076,7 @@ def getloadavg():
     if hasattr(os, "getloadavg"):
         return os.getloadavg()
     else:
-        return _psplatform.getloadavg()  # Windows
+        return _psplatform.getloadavg()
 
 
 # =====================================================================
@@ -2016,7 +2084,7 @@ def getloadavg():
 # =====================================================================
 
 
-def virtual_memory():
+def virtual_memory() -> svmem:
     """Return statistics about system memory usage as a namedtuple
     including the following fields, expressed in bytes:
 
@@ -2075,7 +2143,7 @@ def virtual_memory():
     return ret
 
 
-def swap_memory():
+def swap_memory() -> sswap:
     """Return system swap memory statistics as a namedtuple including
     the following fields:
 
@@ -2096,7 +2164,7 @@ def swap_memory():
 # =====================================================================
 
 
-def disk_usage(path):
+def disk_usage(path: str) -> sdiskusage:
     """Return disk usage statistics about the given *path* as a
     namedtuple including total, used and free space expressed in bytes
     plus the percentage usage.
@@ -2104,7 +2172,7 @@ def disk_usage(path):
     return _psplatform.disk_usage(path)
 
 
-def disk_partitions(all=False):
+def disk_partitions(all: bool = False) -> list[sdiskpart]:
     """Return mounted partitions as a list of
     (device, mountpoint, fstype, opts) namedtuple.
     'opts' field is a raw string separated by commas indicating mount
@@ -2116,7 +2184,9 @@ def disk_partitions(all=False):
     return _psplatform.disk_partitions(all)
 
 
-def disk_io_counters(perdisk=False, nowrap=True):
+def disk_io_counters(
+    perdisk: bool = False, nowrap: bool = True
+) -> sdiskio | dict[str, sdiskio]:
     """Return system disk I/O statistics as a namedtuple including
     the following fields:
 
@@ -2173,7 +2243,9 @@ disk_io_counters.cache_clear.__doc__ = "Clears nowrap argument cache"
 # =====================================================================
 
 
-def net_io_counters(pernic=False, nowrap=True):
+def net_io_counters(
+    pernic: bool = False, nowrap: bool = True
+) -> snetio | dict[str, snetio] | None:
     """Return network I/O statistics as a namedtuple including
     the following fields:
 
@@ -2218,7 +2290,7 @@ net_io_counters.cache_clear = functools.partial(
 net_io_counters.cache_clear.__doc__ = "Clears nowrap argument cache"
 
 
-def net_connections(kind='inet'):
+def net_connections(kind: str = 'inet') -> list[sconn]:
     """Return system-wide socket connections as a list of
     (fd, family, type, laddr, raddr, status, pid) namedtuples.
     In case of limited privileges 'fd' and 'pid' may be set to -1
@@ -2248,7 +2320,7 @@ def net_connections(kind='inet'):
     return _psplatform.net_connections(kind)
 
 
-def net_if_addrs():
+def net_if_addrs() -> dict[str, list[snicaddr]]:
     """Return the addresses associated to each NIC (network interface
     card) installed on the system as a dictionary whose keys are the
     NIC names and value is a list of namedtuples for each address
@@ -2309,7 +2381,7 @@ def net_if_addrs():
     return dict(ret)
 
 
-def net_if_stats():
+def net_if_stats() -> dict[str, snicstats]:
     """Return information about each NIC (network interface card)
     installed on the system as a dictionary whose keys are the
     NIC names and value is a namedtuple with the following fields:
@@ -2332,7 +2404,9 @@ def net_if_stats():
 # Linux, macOS
 if hasattr(_psplatform, "sensors_temperatures"):
 
-    def sensors_temperatures(fahrenheit=False):
+    def sensors_temperatures(
+        fahrenheit: bool = False,
+    ) -> dict[str, list[shwtemp]]:
         """Return hardware temperatures. Each entry is a namedtuple
         representing a certain hardware sensor (it may be a CPU, an
         hard disk or something else, depending on the OS and its
@@ -2370,7 +2444,7 @@ if hasattr(_psplatform, "sensors_temperatures"):
 # Linux
 if hasattr(_psplatform, "sensors_fans"):
 
-    def sensors_fans():
+    def sensors_fans() -> dict[str, list[sfan]]:
         """Return fans speed. Each entry is a namedtuple
         representing a certain hardware sensor.
         All speed are expressed in RPM (rounds per minute).
@@ -2383,7 +2457,7 @@ if hasattr(_psplatform, "sensors_fans"):
 # Linux, Windows, FreeBSD, macOS
 if hasattr(_psplatform, "sensors_battery"):
 
-    def sensors_battery():
+    def sensors_battery() -> sbattery | None:
         """Return battery information. If no battery is installed
         returns None.
 
@@ -2403,7 +2477,7 @@ if hasattr(_psplatform, "sensors_battery"):
 # =====================================================================
 
 
-def boot_time():
+def boot_time() -> float:
     """Return the system boot time expressed in seconds since the epoch
     (seconds since January 1, 1970, at midnight UTC). The returned
     value is based on the system clock, which means it may be affected
@@ -2413,7 +2487,7 @@ def boot_time():
     return _psplatform.boot_time()
 
 
-def users():
+def users() -> list[suser]:
     """Return users currently connected on the system as a list of
     namedtuples including the following fields.
 
@@ -2433,13 +2507,13 @@ def users():
 
 if WINDOWS:
 
-    def win_service_iter():
+    def win_service_iter() -> Iterator[WindowsService]:
         """Return a generator yielding a WindowsService instance for all
         Windows services installed.
         """
         return _psplatform.win_service_iter()
 
-    def win_service_get(name):
+    def win_service_get(name) -> WindowsService:
         """Get a Windows service by *name*.
         Raise NoSuchProcess if no service with such name exists.
         """
@@ -2454,7 +2528,7 @@ if WINDOWS:
 # Linux + glibc, Windows, macOS, FreeBSD, NetBSD
 if hasattr(_psplatform, "heap_info"):
 
-    def heap_info():
+    def heap_info() -> pheap:
         """Return low-level heap statistics from the C heap allocator
         (glibc).
 
@@ -2470,7 +2544,7 @@ if hasattr(_psplatform, "heap_info"):
         """
         return _ntp.pheap(*_psplatform.heap_info())
 
-    def heap_trim():
+    def heap_trim() -> None:
         """Request that the underlying allocator free any unused memory
         it's holding in the heap (typically small `malloc()`
         allocations).